requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-allauth[socialaccount]>=0.57.0
psycopg2-binary>=2.9
gunicorn>=21.0
python-dotenv>=1.0
requests>=2.31
PyJWT>=2.8
cryptography>=41.0
whitenoise>=6.6
dj-database-url>=2.1
